backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.1
sqlalchemy==2.0.23
aiosqlite==0.19.0 